Fixed an -m64 issue for radar 4211430.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/apple-local-200502-branch@103529 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/g++.dg/asm-static-entry-1.C
@@ -5,7 +5,7 @@
 
 extern "C" void abort();
 
-typedef unsigned int UInt32;
+typedef unsigned long UInt32;
 
 #define FUNCPTR_TO_UINT(f) ((UInt32)(&(f)))
